Make profile page a server component, delete reports with server action
